refactor(pipeline): split unified_pipeline.py into modular sub-package (#310)
Split the 5050-line monolithic unified_pipeline.py into 6 focused modules
under a unified_pipeline/ package:
- _helpers.py (440 lines): shared utilities, dotenv, logger
- invoke_callables.py (3079 lines): 35 _invoke_* + 4 private helpers
- state_writers.py (760 lines): 35 _write_* + CAPABILITY_STATE_WRITERS
- registry_setup.py (528 lines): setup_registry + _declare_tweety_slots
- workflow_builders.py (728 lines): 12 build_* + catalog + run_unified_analysis
- __init__.py (~125 lines): re-exports all symbols for backward compat

All 161 existing tests pass unchanged. Zero import changes needed in any
of the 34+ external files that import from unified_pipeline.
