+++ argumentation_analysis/orchestration/unified_pipeline/invoke_callables.py
@@ -1,68 +1,28 @@
+"""Invoke callables for registered pipeline components.
+
+Each callable: async (input_text: str, context: Dict[str, Any]) -> Any
+Registered as invoke= parameter in CapabilityRegistry.
 """
-Unified pipeline wiring all integrated student components via CapabilityRegistry.
-
-Provides:
-- setup_registry(): register all available components (agents, services, adapters)
-- Pre-built workflow definitions (light, standard, full, custom)
-- run_unified_analysis(): convenience function for end-to-end analysis
-
-This module ties together the Lego architecture built in Phases 0-4:
-- CapabilityRegistry (core/capability_registry.py)
-- WorkflowDSL (orchestration/workflow_dsl.py)
-- All integrated student project components
-"""
-
 import asyncio
 import json
 import logging
 import os
-from typing import Dict, Any, Optional, List, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
-from argumentation_analysis.core.capability_registry import (
-    CapabilityRegistry,
-    ServiceDiscovery,
-    ComponentType,
+from argumentation_analysis.orchestration.unified_pipeline._helpers import (
+    _aggregate_virtue_scores,
+    _enrich_ranking_with_justification,
+    _evaluate_counter_arguments,
+    _extract_arguments_from_context,
+    _generate_attacks_from_args,
+    _get_openai_client,
+    _normalize_fallacies_with_quotes,
+    _normalize_items_with_quotes,
+    _python_ranking_fallback,
+    logger,
 )
-from argumentation_analysis.orchestration.workflow_dsl import (
-    WorkflowBuilder,
-    WorkflowDefinition,
-    WorkflowExecutor,
-    PhaseResult,
-    PhaseStatus,
-)
-
-# Ensure .env is loaded so OPENAI_API_KEY is available for all invoke callables
-try:
-    from dotenv import load_dotenv
-
-    load_dotenv()
-except ImportError:
-    pass  # dotenv optional in CI environments
-
-logger = logging.getLogger("UnifiedPipeline")
-
-
-def _get_openai_client() -> Tuple[Any, str]:
-    """Create an AsyncOpenAI client from environment variables.
-
-    Returns (client, model_id) or (None, "") if API key unavailable.
-    """
-    api_key = os.environ.get("OPENAI_API_KEY", "")
-    if not api_key:
-        return None, ""
-    try:
-        from openai import AsyncOpenAI
-
-        base_url = os.environ.get("OPENAI_BASE_URL", "https://api.openai.com/v1")
-        model_id = os.environ.get("OPENAI_CHAT_MODEL_ID", "gpt-5-mini")
-        return AsyncOpenAI(api_key=api_key, base_url=base_url), model_id
-    except ImportError:
-        return None, ""
-
 
 # --- Invoke callables for registered components ---
-# Each callable: async (input_text: str, context: Dict[str, Any]) -> Any
-
 
 async def _invoke_quality_evaluator(input_text: str, context: Dict[str, Any]) -> Dict:
     """Invoke 9-virtue argument quality evaluator on each extracted argument.
@@ -3117,1934 +3077,3 @@ async def _invoke_formal_synthesis(input_text: str, context: Dict[str, Any]) -> 
 # Each writer extracts relevant data from phase output and writes to
 # UnifiedAnalysisState via its typed add_*() methods.
 # Writers are defensive: guard with isinstance checks and .get() everywhere.
-
-
-def _write_quality_to_state(output, state, ctx) -> None:
-    """Write quality evaluator results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-
-    # New format: per-argument scores
-    per_arg = output.get("per_argument_scores", {})
-    if isinstance(per_arg, dict) and per_arg:
-        for arg_id, result in per_arg.items():
-            if not isinstance(result, dict):
-                continue
-            scores = result.get("scores_par_vertu", {})
-            if not isinstance(scores, dict):
-                scores = {}
-            overall = result.get("note_finale", 0.0)
-            llm_assessment = result.get("llm_assessment")  # (#290)
-            if isinstance(overall, (int, float)) and (scores or overall > 0):
-                state.add_quality_score(
-                    str(arg_id), scores, float(overall),
-                    llm_assessment=llm_assessment,
-                )
-        return
-
-    # Legacy format: single evaluation
-    arg_id = ctx.get("current_arg_id", "arg_input")
-    scores = output.get("scores_par_vertu", {})
-    if not scores:
-        scores = {
-            k: v
-            for k, v in output.items()
-            if k
-            not in (
-                "note_finale",
-                "note_moyenne",
-                "scores_par_vertu",
-                "rapport_detaille",
-                "per_argument_scores",
-                "aggregate_score",
-                "arguments_evaluated",
-            )
-            and isinstance(v, (int, float))
-        }
-    overall = output.get("note_finale", 0.0)
-    if isinstance(overall, (int, float)) and (scores or overall > 0):
-        state.add_quality_score(arg_id, scores, float(overall))
-
-
-def _write_counter_argument_to_state(output, state, ctx) -> None:
-    """Write counter-argument results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    strength_map = {"weak": 0.3, "moderate": 0.6, "strong": 0.9}
-
-    # Write ALL LLM-generated counter-arguments
-    llm_cas = output.get("llm_counter_arguments", [])
-    if isinstance(llm_cas, list) and llm_cas:
-        for llm_ca in llm_cas:
-            if not isinstance(llm_ca, dict) or not llm_ca.get("counter_argument"):
-                continue
-            target = str(llm_ca.get("target_argument", ""))[:200]
-            counter_text = str(llm_ca.get("counter_argument", ""))
-            strategy_name = str(llm_ca.get("strategy_used", "unknown"))
-            # (#294) Use evaluation score if available, else fallback to strength map
-            evaluation = llm_ca.get("evaluation", {})
-            if isinstance(evaluation, dict) and "overall_score" in evaluation:
-                score = float(evaluation["overall_score"])
-            else:
-                score = strength_map.get(str(llm_ca.get("strength", "")).lower(), 0.5)
-            state.add_counter_argument(target, counter_text, strategy_name, score)
-        return
-
-    # Backward compat: single LLM counter-argument
-    llm_ca = output.get("llm_counter_argument")
-    if isinstance(llm_ca, dict) and llm_ca.get("counter_argument"):
-        target = str(llm_ca.get("target_argument", ""))[:200]
-        counter_text = str(llm_ca.get("counter_argument", ""))
-        strategy_name = str(llm_ca.get("strategy_used", "unknown"))
-        score = strength_map.get(str(llm_ca.get("strength", "")).lower(), 0.5)
-        state.add_counter_argument(target, counter_text, strategy_name, score)
-        return
-
-    # Fallback to heuristic plugin output
-    parsed = output.get("parsed_argument", {})
-    strategy = output.get("suggested_strategy", {})
-    if not isinstance(parsed, dict):
-        parsed = {}
-    if not isinstance(strategy, dict):
-        strategy = {}
-    original = str(parsed.get("premise", ctx.get("input_data", "")))[:200]
-    strategy_name = str(strategy.get("strategy_name", "unknown"))
-    score = strategy.get("confidence", 0.0)
-    if not isinstance(score, (int, float)):
-        score = 0.0
-    state.add_counter_argument(original, strategy_name, strategy_name, float(score))
-
-
-def _write_jtms_to_state(output, state, ctx) -> None:
-    """Write JTMS results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    beliefs = output.get("beliefs", {})
-    if not isinstance(beliefs, dict):
-        return
-    for name, belief_data in beliefs.items():
-        if isinstance(belief_data, dict):
-            valid = belief_data.get("valid")
-            justifications = belief_data.get("justifications", [])
-        else:
-            # Legacy format: belief_data is a string like "True"/"False"/"None"
-            valid_str = str(belief_data)
-            valid = (
-                True
-                if valid_str == "True"
-                else (False if valid_str == "False" else None)
-            )
-            justifications = []
-        if not isinstance(justifications, list):
-            justifications = []
-        state.add_jtms_belief(str(name), valid, justifications=justifications)
-
-
-def _write_atms_to_state(output, state, ctx) -> None:
-    """Write ATMS assumption-based reasoning results to UnifiedAnalysisState (#292).
-
-    Stores each node's environment info as a JTMS belief for compatibility.
-    """
-    if not output or not isinstance(output, dict):
-        return
-    environments = output.get("environments", {})
-    if not isinstance(environments, dict):
-        return
-    for name, env_data in environments.items():
-        if not isinstance(env_data, dict):
-            continue
-        is_assumption = env_data.get("is_assumption", False)
-        env_list = env_data.get("environments", [])
-        # For ATMS, "valid" = has at least one consistent environment
-        valid = len(env_list) > 0
-        justifications = [
-            f"assumption_env:{sorted(e)}" for e in env_list[:5]
-        ]
-        state.add_jtms_belief(
-            f"ATMS:{name}", valid, justifications=justifications
-        )
-    # Store summary metadata
-    state.add_jtms_belief(
-        "ATMS:summary",
-        True,
-        justifications=[
-            f"assumptions={output.get('assumption_count', 0)}",
-            f"nodes={output.get('node_count', 0)}",
-            f"consistent_derivations={len(output.get('consistent_derivations', []))}",
-            f"contradictions={'yes' if output.get('has_contradictions') else 'no'}",
-        ],
-    )
-
-
-def _write_debate_to_state(output, state, ctx) -> None:
-    """Write debate analysis results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    topic = str(ctx.get("input_data", ""))[:100]
-    winner = output.get("winner")
-    # Build exchanges from LLM debate assessment if available
-    exchanges = []
-    llm_debate = output.get("llm_debate_assessment")
-    if isinstance(llm_debate, dict):
-        key_exchanges = llm_debate.get("key_exchanges", [])
-        if isinstance(key_exchanges, list):
-            for ex in key_exchanges:
-                if isinstance(ex, dict):
-                    exchanges.append(
-                        {
-                            "point": str(ex.get("point", "")),
-                            "rebuttal": str(ex.get("rebuttal", "")),
-                        }
-                    )
-    state.add_debate_transcript(
-        topic=topic,
-        exchanges=exchanges,
-        winner=str(winner) if winner is not None else None,
-    )
-
-
-def _write_governance_to_state(output, state, ctx) -> None:
-    """Write governance results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-
-    llm_gov = output.get("llm_governance_assessment", {})
-    if not isinstance(llm_gov, dict):
-        llm_gov = {}
-
-    # Build scores from stakeholder analysis or conflicts
-    scores = {}
-    stakeholders = llm_gov.get("stakeholder_analysis", [])
-    if isinstance(stakeholders, list):
-        for s in stakeholders:
-            if isinstance(s, dict):
-                agent = str(s.get("agent", "unknown"))
-                influence = float(s.get("influence", 0.0))
-                scores[agent] = influence
-
-    # Fallback: use available methods as score keys if no stakeholders
-    if not scores:
-        methods = output.get("available_methods", [])
-        if isinstance(methods, list) and methods:
-            scores = {str(m): 0.0 for m in methods}
-
-    # If no scores at all (no methods, no stakeholders, no LLM), skip
-    has_conflicts = bool(output.get("conflicts"))
-    has_llm = bool(llm_gov)
-    if not scores and not has_conflicts and not has_llm:
-        return
-
-    recommended = output.get("recommended_method") or llm_gov.get(
-        "recommended_method", "majority"
-    )
-
-    # Determine winner from vote result, LLM assessment, or conflict resolution
-    winner = "N/A"
-    vote_result = output.get("vote_result", {})
-    if isinstance(vote_result, dict) and vote_result.get("winner"):
-        winner = str(vote_result["winner"])
-        # (#294) Merge Copeland scores into scores dict
-        copeland_scores = vote_result.get("copeland_scores", {})
-        if isinstance(copeland_scores, dict):
-            for agent, cscore in copeland_scores.items():
-                scores[str(agent)] = float(cscore)
-    elif llm_gov.get("recommended_resolution"):
-        winner = str(llm_gov["recommended_resolution"])
-    elif output.get("resolutions"):
-        resolutions = output["resolutions"]
-        if isinstance(resolutions, list) and resolutions:
-            winner = str(resolutions[0].get("resolution_type", "N/A"))
-
-    state.add_governance_decision(
-        method=str(recommended),
-        winner=winner,
-        scores=scores,
-    )
-
-
-def _write_camembert_to_state(output, state, ctx) -> None:
-    """Write CamemBERT neural fallacy results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    detections = output.get("detections", [])
-    if not isinstance(detections, list):
-        return
-    for det in detections:
-        if not isinstance(det, dict):
-            continue
-        state.add_neural_fallacy_score(
-            text_segment=str(det.get("text", "")),
-            label=str(det.get("label", "unknown")),
-            confidence=float(det.get("confidence", 0.0)),
-        )
-
-
-def _write_hierarchical_fallacy_to_state(output, state, ctx) -> None:
-    """Write hierarchical taxonomy-guided fallacy results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    fallacies = output.get("fallacies", [])
-    if not isinstance(fallacies, list):
-        return
-    for f in fallacies:
-        if not isinstance(f, dict):
-            continue
-        fallacy_type = f.get("type", f.get("fallacy_type", "unknown"))
-        justification = f.get("explanation", "")
-        taxonomy_pk = f.get("taxonomy_pk", "")
-        confidence = f.get("confidence", 0.0)
-        trace = f.get("navigation_trace", [])
-        full_justification = justification
-        if taxonomy_pk:
-            full_justification += f" [taxonomy:{taxonomy_pk}]"
-        if confidence:
-            full_justification += f" [confidence:{confidence:.2f}]"
-        if trace:
-            full_justification += f" [trace:{'>'.join(trace)}]"
-        state.add_fallacy(fallacy_type=fallacy_type, justification=full_justification)
-
-
-def _write_semantic_index_to_state(output, state, ctx) -> None:
-    """Write semantic index results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    results = output.get("results", [])
-    if not isinstance(results, list):
-        return
-    query = str(ctx.get("input_data", ""))
-    for r in results:
-        if not isinstance(r, dict):
-            continue
-        state.add_semantic_index_ref(
-            query=query,
-            document_id=str(r.get("id", "unknown")),
-            score=float(r.get("score", 0.0)),
-            snippet=r.get("snippet"),
-        )
-
-
-def _write_speech_to_state(output, state, ctx) -> None:
-    """Write speech transcription results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    segments = output.get("segments", [])
-    if not isinstance(segments, list):
-        return
-    for seg in segments:
-        if not isinstance(seg, dict):
-            continue
-        state.add_transcription_segment(
-            start_time=float(seg.get("start", 0.0)),
-            end_time=float(seg.get("end", 0.0)),
-            text=str(seg.get("text", "")),
-            speaker=seg.get("speaker"),
-        )
-
-
-def _write_ranking_to_state(output, state, ctx) -> None:
-    """Write ranking semantics results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    method = str(output.get("method", "unknown"))
-    arguments = output.get("arguments", [])
-    comparisons = output.get("comparisons", [])
-    if not isinstance(arguments, list):
-        arguments = []
-    if not isinstance(comparisons, list):
-        comparisons = []
-    state.add_ranking_result(method, arguments, comparisons)
-
-
-def _write_aspic_to_state(output, state, ctx) -> None:
-    """Write ASPIC+ analysis results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    reasoner_type = str(output.get("reasoner_type", "simple"))
-    extensions = output.get("extensions", [])
-    statistics = output.get("statistics", {})
-    if not isinstance(extensions, list):
-        extensions = []
-    if not isinstance(statistics, dict):
-        statistics = {}
-    state.add_aspic_result(reasoner_type, extensions, statistics)
-
-
-def _write_belief_revision_to_state(output, state, ctx) -> None:
-    """Write belief revision results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    method = str(output.get("method", "dalal"))
-    original = output.get("original", [])
-    revised = output.get("revised", [])
-    if not isinstance(original, list):
-        original = []
-    if not isinstance(revised, list):
-        revised = []
-    state.add_belief_revision_result(method, original, revised)
-
-
-def _write_dialogue_to_state(output, state, ctx) -> None:
-    """Write dialogue protocol results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    topic = str(output.get("topic", ""))
-    outcome = str(output.get("outcome", "unknown"))
-    trace = output.get("dialogue_trace", [])
-    if not isinstance(trace, list):
-        trace = []
-    state.add_dialogue_result(topic, outcome, trace)
-
-
-def _write_probabilistic_to_state(output, state, ctx) -> None:
-    """Write probabilistic argumentation results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    arguments = output.get("arguments", [])
-    acceptance = output.get("acceptance_probabilities", {})
-    if not isinstance(arguments, list):
-        arguments = []
-    if not isinstance(acceptance, dict):
-        acceptance = {}
-    state.add_probabilistic_result(arguments, acceptance)
-
-
-def _write_bipolar_to_state(output, state, ctx) -> None:
-    """Write bipolar argumentation results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    fw_type = str(output.get("framework_type", "necessity"))
-    arguments = output.get("arguments", [])
-    supports = output.get("supports", [])
-    if not isinstance(arguments, list):
-        arguments = []
-    if not isinstance(supports, list):
-        supports = []
-    state.add_bipolar_result(fw_type, arguments, supports)
-
-
-def _write_aba_to_state(output, state, ctx) -> None:
-    """Write ABA reasoning results to UnifiedAnalysisState (stored as Dung framework)."""
-    if not output or not isinstance(output, dict):
-        return
-    assumptions = output.get("assumptions", [])
-    extensions = output.get("extensions", [])
-    if not isinstance(assumptions, list):
-        assumptions = []
-    state.add_dung_framework(
-        name=f"aba_{output.get('semantics', 'preferred')}",
-        arguments=assumptions,
-        attacks=[],
-        extensions={"aba_extensions": extensions},
-    )
-
-
-def _write_adf_to_state(output, state, ctx) -> None:
-    """Write ADF reasoning results to UnifiedAnalysisState (stored as Dung framework)."""
-    if not output or not isinstance(output, dict):
-        return
-    statements = output.get("statements", [])
-    models = output.get("models", output.get("extensions", []))
-    if not isinstance(statements, list):
-        statements = []
-    state.add_dung_framework(
-        name=f"adf_{output.get('semantics', 'grounded')}",
-        arguments=statements,
-        attacks=[],
-        extensions={"adf_models": models},
-    )
-
-
-def _write_fact_extraction_to_state(output, state, ctx) -> None:
-    """Write fact extraction results to state (populates extracts + base fields)."""
-    if not output or not isinstance(output, dict):
-        return
-    # Write claims to extracts (with source quotes when available)
-    claims = output.get("claims", [])
-    if isinstance(claims, list):
-        for claim in claims:
-            if isinstance(claim, dict):
-                text = claim.get("text", "").strip()
-                quote = claim.get("source_quote", "")
-                if text:
-                    entry = {"type": "claim", "content": text}
-                    if quote:
-                        entry["source_quote"] = quote
-                    state.extracts.append(entry)
-            elif isinstance(claim, str) and claim.strip():
-                state.extracts.append({"type": "claim", "content": claim.strip()})
-    # Populate base identified_arguments from LLM extraction
-    arguments = output.get("arguments", [])
-    if isinstance(arguments, list):
-        for arg in arguments:
-            if isinstance(arg, dict):
-                text = arg.get("text", "").strip()
-                quote = arg.get("source_quote", "")
-                if text:
-                    arg_text = text
-                    if quote:
-                        arg_text = f'{text} [quote: "{quote[:100]}"]'
-                    state.add_argument(arg_text)
-            elif isinstance(arg, str) and arg.strip():
-                state.add_argument(arg.strip())
-    # NOTE: Fallacy detection removed from fact_extraction (issue #179).
-    # Fallacies are the sole responsibility of hierarchical_fallacy_detection,
-    # which uses deep taxonomy navigation for precise identification.
-    # Set summary as analysis task if present
-    summary = output.get("summary", "")
-    if summary and isinstance(summary, str):
-        state.add_task(f"Fact extraction: {summary[:200]}")
-
-
-def _write_propositional_to_state(output, state, ctx) -> None:
-    """Write propositional logic analysis results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    formulas = output.get("formulas", [])
-    satisfiable = output.get("satisfiable", False)
-    model = output.get("model", {})
-    if not isinstance(formulas, list):
-        formulas = []
-    if not isinstance(model, dict):
-        model = {}
-    state.add_propositional_analysis_result(formulas, bool(satisfiable), model)
-
-
-def _write_fol_to_state(output, state, ctx) -> None:
-    """Write FOL reasoning results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    formulas = output.get("formulas", [])
-    consistent = output.get("consistent", False)
-    inferences = output.get("inferences", [])
-    confidence = output.get("confidence", 0.0)
-    if not isinstance(formulas, list):
-        formulas = []
-    if not isinstance(inferences, list):
-        inferences = []
-    if not isinstance(confidence, (int, float)):
-        confidence = 0.0
-    state.add_fol_analysis_result(
-        formulas, bool(consistent), inferences, float(confidence)
-    )
-
-
-def _write_modal_to_state(output, state, ctx) -> None:
-    """Write modal logic analysis results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    formulas = output.get("formulas", [])
-    valid = output.get("valid", False)
-    modalities = output.get("modalities", [])
-    if not isinstance(formulas, list):
-        formulas = []
-    if not isinstance(modalities, list):
-        modalities = []
-    state.add_modal_analysis_result(formulas, bool(valid), modalities)
-
-
-def _write_nl_to_logic_to_state(output, state, ctx) -> None:
-    """Write NL-to-formal-logic translation results to UnifiedAnalysisState (#173)."""
-    if not output or not isinstance(output, dict):
-        return
-    translations = output.get("translations", [])
-    if not isinstance(translations, list):
-        return
-    for t in translations:
-        if isinstance(t, dict) and t.get("formula"):
-            state.add_nl_to_logic_translation(
-                original_text=t.get("original_text", ""),
-                formula=t.get("formula", ""),
-                logic_type=t.get("logic_type", "propositional"),
-                is_valid=bool(t.get("is_valid", False)),
-                variables=t.get("variables", {}),
-                confidence=float(t.get("confidence", 0.0)),
-            )
-
-
-def _write_dung_extensions_to_state(output, state, ctx) -> None:
-    """Write Dung extension computation results to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    semantics = str(output.get("semantics", "preferred"))
-    extensions = output.get("extensions", {})
-    statistics = output.get("statistics", {})
-    arguments = []
-    if isinstance(statistics, dict):
-        arguments = [f"arg_{i}" for i in range(statistics.get("arguments_count", 0))]
-    state.add_dung_framework(
-        name=f"verification_{semantics}",
-        arguments=arguments,
-        attacks=[],
-        extensions=extensions if isinstance(extensions, dict) else {},
-    )
-
-
-def _write_formal_synthesis_to_state(output, state, ctx) -> None:
-    """Write formal synthesis report to UnifiedAnalysisState."""
-    if not output or not isinstance(output, dict):
-        return
-    summary = str(output.get("summary", ""))
-    phase_results = output.get("phase_results", {})
-    overall_validity = output.get("overall_validity", 0.0)
-    if not isinstance(phase_results, dict):
-        phase_results = {}
-    if not isinstance(overall_validity, (int, float)):
-        overall_validity = 0.0
-    state.add_formal_synthesis_report(summary, phase_results, float(overall_validity))
-
-
-def _write_dl_to_state(output, state, ctx) -> None:
-    """Write Description Logic results to UnifiedAnalysisState (#86)."""
-    if not output or not isinstance(output, dict):
-        return
-    consistent = output.get("consistent", False)
-    message = str(output.get("message", ""))
-    state.add_fol_analysis_result(
-        formulas=[f"DL: {message}"],
-        consistent=bool(consistent),
-        inferences=[],
-        confidence=1.0 if consistent else 0.0,
-    )
-
-
-def _write_cl_to_state(output, state, ctx) -> None:
-    """Write Conditional Logic results to UnifiedAnalysisState (#86)."""
-    if not output or not isinstance(output, dict):
-        return
-    entailed = output.get("entailed", False)
-    message = str(output.get("message", ""))
-    num = output.get("num_conditionals", 0)
-    state.add_propositional_analysis_result(
-        formulas=[f"CL({num} conditionals): {message}"],
-        satisfiable=bool(entailed),
-        model={},
-    )
-
-
-def _write_sat_to_state(output, state, ctx) -> None:
-    """Write SAT solver results to UnifiedAnalysisState (#86)."""
-    if not output or not isinstance(output, dict):
-        return
-    is_sat = output.get("satisfiable", False)
-    model = output.get("model") or {}
-    mode = output.get("mode", "solve")
-    if mode == "mus":
-        mus_count = output.get("mus_count", 0)
-        state.add_propositional_analysis_result(
-            formulas=[f"SAT/MUS: {mus_count} minimal unsatisfiable subsets"],
-            satisfiable=False,
-            model={},
-        )
-    else:
-        state.add_propositional_analysis_result(
-            formulas=[f"SAT: {'SAT' if is_sat else 'UNSAT'}"],
-            satisfiable=bool(is_sat),
-            model=model if isinstance(model, dict) else {},
-        )
-
-
-def _write_setaf_to_state(output, state, ctx) -> None:
-    """Write SetAF results to UnifiedAnalysisState (#87)."""
-    if not output or not isinstance(output, dict):
-        return
-    state.add_dung_framework(
-        name=f"setaf_{output.get('semantics', 'grounded')}",
-        arguments=output.get("arguments", []),
-        attacks=[],  # set attacks don't map to binary attacks
-        extensions={"setaf_extensions": output.get("extensions", [])},
-    )
-
-
-def _write_weighted_to_state(output, state, ctx) -> None:
-    """Write Weighted AF results to UnifiedAnalysisState (#87)."""
-    if not output or not isinstance(output, dict):
-        return
-    state.add_dung_framework(
-        name=f"weighted_{output.get('semantics', 'grounded')}",
-        arguments=output.get("arguments", []),
-        attacks=[
-            [a.get("source", ""), a.get("target", "")]
-            for a in output.get("attacks", [])
-            if isinstance(a, dict)
-        ],
-        extensions={"weighted_extensions": output.get("extensions", [])},
-    )
-
-
-def _write_social_to_state(output, state, ctx) -> None:
-    """Write Social AF results to UnifiedAnalysisState (#87)."""
-    if not output or not isinstance(output, dict):
-        return
-    ranking = output.get("ranking", [])
-    scores = output.get("scores", {})
-    state.add_dung_framework(
-        name="social_af",
-        arguments=output.get("arguments", []),
-        attacks=output.get("attacks", []),
-        extensions={"social_ranking": ranking, "social_scores": scores},
-    )
-
-
-def _write_eaf_to_state(output, state, ctx) -> None:
-    """Write EAF results to UnifiedAnalysisState (#88)."""
-    if not output or not isinstance(output, dict):
-        return
-    state.add_dung_framework(
-        name=f"eaf_{output.get('semantics', 'grounded')}",
-        arguments=output.get("arguments", []),
-        attacks=[a for a in output.get("attacks", []) if isinstance(a, list)],
-        extensions={"eaf_extensions": output.get("extensions", [])},
-    )
-
-
-def _write_delp_to_state(output, state, ctx) -> None:
-    """Write DeLP results to UnifiedAnalysisState (#89)."""
-    if not output or not isinstance(output, dict):
-        return
-    query_results = output.get("query_results", [])
-    state.add_dung_framework(
-        name="delp_analysis",
-        arguments=[],
-        attacks=[],
-        extensions={"delp_query_results": query_results},
-    )
-
-
-def _write_qbf_to_state(output, state, ctx) -> None:
-    """Write QBF results to UnifiedAnalysisState (#90)."""
-    if not output or not isinstance(output, dict):
-        return
-    state.add_propositional_analysis_result(
-        formulas=[f"QBF: {output.get('formula', '')}"],
-        satisfiable=output.get("valid", False),
-        model={},
-    )
-
-
-def _write_collaborative_analysis_to_state(output, state, ctx) -> None:
-    """Write collaborative multi-agent debate results to state (#175)."""
-    from argumentation_analysis.orchestration.collaborative_debate import (
-        _write_collaborative_to_state,
-    )
-
-    _write_collaborative_to_state(output, state, ctx)
-
-
-CAPABILITY_STATE_WRITERS: Dict[str, Any] = {
-    "argument_quality": _write_quality_to_state,
-    "counter_argument_generation": _write_counter_argument_to_state,
-    "belief_maintenance": _write_jtms_to_state,
-    "assumption_based_reasoning": _write_atms_to_state,
-    "adversarial_debate": _write_debate_to_state,
-    "governance_simulation": _write_governance_to_state,
-    "neural_fallacy_detection": _write_camembert_to_state,
-    "semantic_indexing": _write_semantic_index_to_state,
-    "speech_transcription": _write_speech_to_state,
-    "ranking_semantics": _write_ranking_to_state,
-    "aspic_plus_reasoning": _write_aspic_to_state,
-    "belief_revision": _write_belief_revision_to_state,
-    "dialogue_protocols": _write_dialogue_to_state,
-    "probabilistic_argumentation": _write_probabilistic_to_state,
-    "bipolar_argumentation": _write_bipolar_to_state,
-    "aba_reasoning": _write_aba_to_state,
-    "adf_reasoning": _write_adf_to_state,
-    "fact_extraction": _write_fact_extraction_to_state,
-    "propositional_logic": _write_propositional_to_state,
-    "fol_reasoning": _write_fol_to_state,
-    "modal_logic": _write_modal_to_state,
-    "dung_extensions": _write_dung_extensions_to_state,
-    "formal_synthesis": _write_formal_synthesis_to_state,
-    "hierarchical_fallacy_detection": _write_hierarchical_fallacy_to_state,
-    "description_logic": _write_dl_to_state,
-    "conditional_logic": _write_cl_to_state,
-    "sat_solving": _write_sat_to_state,
-    "setaf_reasoning": _write_setaf_to_state,
-    "weighted_argumentation": _write_weighted_to_state,
-    "social_argumentation": _write_social_to_state,
-    "epistemic_argumentation": _write_eaf_to_state,
-    "defeasible_logic": _write_delp_to_state,
-    "qbf_reasoning": _write_qbf_to_state,
-    "collaborative_analysis": _write_collaborative_analysis_to_state,
-    "nl_to_logic_translation": _write_nl_to_logic_to_state,
-}
-
-
-def setup_registry(
-    include_optional: bool = True,
-) -> CapabilityRegistry:
-    """
-    Create and populate a CapabilityRegistry with all available components.
-
-    Registers agents, services, and adapters from the integrated student projects.
-    Each registration is wrapped in try/except so missing optional dependencies
-    don't prevent the registry from being created.
-
-    Args:
-        include_optional: If True, attempt to register components with optional
-                         dependencies (CamemBERT, Dung/JVM, speech-to-text).
-
-    Returns:
-        Populated CapabilityRegistry instance.
-    """
-    registry = CapabilityRegistry()
-    registered = []
-    skipped = []
-
-    # --- Core agents (always available) ---
-
-    # Counter-argument generation (2.3.3)
-    try:
-        from argumentation_analysis.agents.core.counter_argument import (
-            register_with_capability_registry as register_counter_arg,
-        )
-
-        register_counter_arg(registry)
-        # Wire invoke callable (registration was created by register_counter_arg)
-        if "counter_argument_agent" in registry._registrations:
-            registry._registrations[
-                "counter_argument_agent"
-            ].invoke = _invoke_counter_argument
-        registered.append("counter_argument_agent")
-    except ImportError as e:
-        skipped.append(("counter_argument_agent", str(e)))
-
-    # Argument quality evaluation (2.3.5)
-    try:
-        from argumentation_analysis.agents.core.quality.quality_evaluator import (
-            ArgumentQualityEvaluator,
-        )
-
-        registry.register_agent(
-            name="quality_evaluator",
-            agent_class=ArgumentQualityEvaluator,
-            capabilities=[
-                "argument_quality",
-                "virtue_detection",
-                "quality_scoring",
-            ],
-            metadata={"description": "9-virtue argument quality evaluator"},
-            invoke=_invoke_quality_evaluator,
-        )
-        registered.append("quality_evaluator")
-    except ImportError as e:
-        skipped.append(("quality_evaluator", str(e)))
-
-    # Debate agent (1.2.7)
-    try:
-        from argumentation_analysis.agents.core.debate.debate_agent import (
-            DebateAgent,
-        )
-
-        registry.register_agent(
-            name="debate_agent",
-            agent_class=DebateAgent,
-            capabilities=[
-                "adversarial_debate",
-                "argument_stress_test",
-                "debate_scoring",
-            ],
-            metadata={"description": "Multi-personality adversarial debate agent"},
-            invoke=_invoke_debate_analysis,
-        )
-        registered.append("debate_agent")
-    except ImportError as e:
-        skipped.append(("debate_agent", str(e)))
-
-    # Governance agent (2.1.6)
-    try:
-        from argumentation_analysis.agents.core.governance.governance_agent import (
-            Agent as GovernanceAgent,
-        )
-
-        registry.register_agent(
-            name="governance_agent",
-            agent_class=GovernanceAgent,
-            capabilities=[
-                "governance_simulation",
-                "multi_method_voting",
-                "preference_aggregation",
-            ],
-            metadata={"description": "7-method governance voting agent"},
-            invoke=_invoke_governance,
-        )
-        registered.append("governance_agent")
-    except ImportError as e:
-        skipped.append(("governance_agent", str(e)))
-
-    # --- Core services ---
-
-    # JTMS belief maintenance (1.4.1)
-    try:
-        from argumentation_analysis.services.jtms.jtms_core import JTMS
-
-        registry.register_service(
-            name="jtms_service",
-            service_class=JTMS,
-            capabilities=["belief_maintenance", "truth_maintenance", "jtms_reasoning"],
-            metadata={"description": "Justification-based Truth Maintenance System"},
-            invoke=_invoke_jtms,
-        )
-        registered.append("jtms_service")
-    except ImportError as e:
-        skipped.append(("jtms_service", str(e)))
-
-    # ATMS assumption-based reasoning (1.4.1) (#292)
-    try:
-        from argumentation_analysis.services.jtms.atms_core import ATMS as ATMSCore
-
-        registry.register_service(
-            name="atms_service",
-            service_class=ATMSCore,
-            capabilities=[
-                "assumption_based_reasoning",
-                "environment_tracking",
-                "atms_reasoning",
-            ],
-            metadata={
-                "description": "Assumption-based Truth Maintenance System (ATMS)"
-            },
-            invoke=_invoke_atms,
-        )
-        registered.append("atms_service")
-    except ImportError as e:
-        skipped.append(("atms_service", str(e)))
-
-    # Local LLM service (2.3.6)
-    try:
-        from argumentation_analysis.services.local_llm_service import (
-            LocalLLMService,
-        )
-
-        registry.register_service(
-            name="local_llm_service",
-            service_class=LocalLLMService,
-            capabilities=["local_llm", "chat_completion"],
-            metadata={"description": "OpenAI-compatible local LLM adapter"},
-            invoke=_invoke_local_llm,
-        )
-        registered.append("local_llm_service")
-    except ImportError as e:
-        skipped.append(("local_llm_service", str(e)))
-
-    # --- Optional adapters (may need heavy dependencies) ---
-
-    if include_optional:
-        # Self-hosted LLM fallacy detector (replaces CamemBERT, #297)
-        try:
-            registry.register_service(
-                name="self_hosted_fallacy_detector",
-                capabilities=["neural_fallacy_detection", "fallacy_detection"],
-                metadata={
-                    "description": (
-                        "Self-hosted LLM fallacy detector via function calling "
-                        "(replaces CamemBERT Tier 2.5, #297)"
-                    )
-                },
-                invoke=_invoke_camembert_fallacy,
-            )
-            registered.append("self_hosted_fallacy_detector")
-        except Exception as e:
-            skipped.append(("self_hosted_fallacy_detector", str(e)))
-
-        # Hierarchical taxonomy-guided fallacy detection (#84)
-        try:
-            from argumentation_analysis.plugins.fallacy_workflow_plugin import (
-                FallacyWorkflowPlugin,
-            )
-
-            registry.register_service(
-                name="hierarchical_fallacy_detector",
-                service_class=FallacyWorkflowPlugin,
-                capabilities=[
-                    "hierarchical_fallacy_detection",
-                    "fallacy_detection",
-                ],
-                metadata={
-                    "description": (
-                        "Hierarchical taxonomy-guided fallacy detection "
-                        "with iterative deepening and one-shot fallback (#84)"
-                    )
-                },
-                invoke=_invoke_hierarchical_fallacy,
-            )
-            registered.append("hierarchical_fallacy_detector")
-        except ImportError as e:
-            skipped.append(("hierarchical_fallacy_detector", str(e)))
-
-        # Semantic index service (Arg_Semantic_Index)
-        try:
-            from argumentation_analysis.services.semantic_index_service import (
-                SemanticIndexService,
-            )
-
-            registry.register_service(
-                name="semantic_index_service",
-                service_class=SemanticIndexService,
-                capabilities=["semantic_indexing", "argument_search"],
-                metadata={"description": "Semantic argument indexing service"},
-                invoke=_invoke_semantic_index,
-            )
-            registered.append("semantic_index_service")
-        except ImportError as e:
-            skipped.append(("semantic_index_service", str(e)))
-
-        # Speech transcription service (speech-to-text)
-        try:
-            from argumentation_analysis.services.speech_transcription_service import (
-                SpeechTranscriptionService,
-            )
-
-            registry.register_service(
-                name="speech_transcription_service",
-                service_class=SpeechTranscriptionService,
-                capabilities=["speech_transcription", "speech_to_text"],
-                metadata={"description": "Whisper-based speech transcription service"},
-                invoke=_invoke_speech_transcription,
-            )
-            registered.append("speech_transcription_service")
-        except ImportError as e:
-            skipped.append(("speech_transcription_service", str(e)))
-
-    # --- Declare unfilled slots for future Tweety extensions ---
-    _declare_tweety_slots(registry)
-
-    # --- TweetyLogicPlugin: SK wrapper for all handlers (#91) ---
-    try:
-        from argumentation_analysis.plugins.tweety_logic_plugin import TweetyLogicPlugin
-
-        registry.register_plugin(
-            name="tweety_logic_plugin",
-            plugin_class=TweetyLogicPlugin,
-            capabilities=["tweety_logic"],
-            metadata={
-                "description": (
-                    "SK plugin exposing all Tweety logic handlers as "
-                    "@kernel_function methods for LLM agents (#91)"
-                )
-            },
-        )
-        registered.append("tweety_logic_plugin")
-    except ImportError as e:
-        skipped.append(("tweety_logic_plugin", str(e)))
-
-    # --- Logic agent capabilities (#71 Formal Verification) ---
-    logic_capabilities = [
-        (
-            "fact_extraction_service",
-            ["fact_extraction"],
-            "Heuristic claim extraction from text",
-            _invoke_fact_extraction,
-        ),
-        (
-            "propositional_logic_service",
-            ["propositional_logic"],
-            "Propositional logic analysis via Tweety",
-            _invoke_propositional_logic,
-        ),
-        (
-            "fol_reasoning_service",
-            ["fol_reasoning"],
-            "First-order logic analysis via Tweety",
-            _invoke_fol_reasoning,
-        ),
-        (
-            "modal_logic_service",
-            ["modal_logic"],
-            "Modal logic analysis via Tweety",
-            _invoke_modal_logic,
-        ),
-        (
-            "dung_extensions_service",
-            ["dung_extensions"],
-            "Dung AF extension computation via AFHandler",
-            _invoke_dung_extensions,
-        ),
-        (
-            "formal_synthesis_service",
-            ["formal_synthesis"],
-            "Aggregate formal analysis into unified report",
-            _invoke_formal_synthesis,
-        ),
-        # SAT handler — no JVM dependency, uses PySAT+Z3 (#86)
-        (
-            "sat_handler",
-            ["sat_solving"],
-            "SAT/MaxSAT/MUS solver (PySAT + Z3)",
-            _invoke_sat,
-        ),
-        # NL-to-formal-logic translation (#173)
-        (
-            "nl_to_logic_service",
-            ["nl_to_logic_translation"],
-            "NL→formal logic with LLM + Tweety validation",
-            _invoke_nl_to_logic,
-        ),
-    ]
-    for name, caps, desc, invoke_fn in logic_capabilities:
-        try:
-            registry.register_service(
-                name=name,
-                service_class=type(name, (), {}),
-                capabilities=caps,
-                metadata={"description": desc},
-                invoke=invoke_fn,
-            )
-            registered.append(name)
-        except Exception as e:
-            skipped.append((name, str(e)))
-
-    # --- Collaborative multi-agent debate (#175) ---
-    try:
-        from argumentation_analysis.orchestration.collaborative_debate import (
-            _invoke_collaborative_analysis,
-        )
-
-        registry.register_service(
-            name="collaborative_debate_service",
-            service_class=type("collaborative_debate_service", (), {}),
-            capabilities=["collaborative_analysis"],
-            metadata={
-                "description": (
-                    "Multi-agent collaborative debate with 4 distinct roles "
-                    "(critic, validator, devil's advocate, synthesizer)"
-                )
-            },
-            invoke=_invoke_collaborative_analysis,
-        )
-        registered.append("collaborative_debate_service")
-    except ImportError as e:
-        skipped.append(("collaborative_debate_service", str(e)))
-
-    logger.info(
-        f"Registry setup complete: {len(registered)} registered, "
-        f"{len(skipped)} skipped"
-    )
-    if skipped:
-        logger.debug(f"Skipped components: {[s[0] for s in skipped]}")
-
-    return registry
-
-
-def _declare_tweety_slots(registry: CapabilityRegistry) -> None:
-    """Register Tweety handler capabilities (Track A #55-#62, #85-#86)."""
-    tweety_handlers = [
-        (
-            "ranking_semantics_handler",
-            ["ranking_semantics"],
-            "Qualitative argument ranking (Categoriser, Burden)",
-            _invoke_ranking,
-        ),
-        (
-            "bipolar_handler",
-            ["bipolar_argumentation"],
-            "Bipolar argumentation (support + attack)",
-            _invoke_bipolar,
-        ),
-        (
-            "aba_handler",
-            ["aba_reasoning"],
-            "Assumption-Based Argumentation",
-            _invoke_aba,
-        ),
-        (
-            "adf_handler",
-            ["adf_reasoning"],
-            "Abstract Dialectical Frameworks",
-            _invoke_adf,
-        ),
-        (
-            "aspic_handler",
-            ["aspic_plus_reasoning"],
-            "ASPIC+ structured argumentation",
-            _invoke_aspic,
-        ),
-        (
-            "belief_revision_handler",
-            ["belief_revision"],
-            "Belief dynamics and revision operators",
-            _invoke_belief_revision,
-        ),
-        (
-            "probabilistic_handler",
-            ["probabilistic_argumentation"],
-            "Probabilistic argument acceptance",
-            _invoke_probabilistic,
-        ),
-        (
-            "dialogue_handler",
-            ["dialogue_protocols"],
-            "Agent dialogue and negotiation protocols",
-            _invoke_dialogue,
-        ),
-        # New handlers (#86)
-        (
-            "dl_handler",
-            ["description_logic"],
-            "ALC Description Logic (TBox/ABox reasoning)",
-            _invoke_dl,
-        ),
-        (
-            "cl_handler",
-            ["conditional_logic"],
-            "Conditional Logic (System Z, non-monotonic)",
-            _invoke_cl,
-        ),
-        # AF variants (#87)
-        (
-            "setaf_handler",
-            ["setaf_reasoning"],
-            "Set Argumentation Frameworks (collective attacks)",
-            _invoke_setaf,
-        ),
-        (
-            "weighted_handler",
-            ["weighted_argumentation"],
-            "Weighted Argumentation Frameworks (attack weights)",
-            _invoke_weighted,
-        ),
-        (
-            "social_handler",
-            ["social_argumentation"],
-            "Social Abstract Argumentation (voting + attacks)",
-            _invoke_social,
-        ),
-        # New handlers (#88, #89, #90)
-        (
-            "eaf_handler",
-            ["epistemic_argumentation"],
-            "Epistemic AF (belief-aware multi-agent argumentation)",
-            _invoke_eaf,
-        ),
-        (
-            "delp_handler",
-            ["defeasible_logic"],
-            "Defeasible Logic Programming (dialectical trees)",
-            _invoke_delp,
-        ),
-        (
-            "qbf_handler",
-            ["qbf_reasoning"],
-            "Quantified Boolean Formulas (∀/∃ over PL)",
-            _invoke_qbf,
-        ),
-    ]
-    for name, caps, desc, invoke_fn in tweety_handlers:
-        try:
-            registry.register_service(
-                name=name,
-                service_class=type(name, (), {}),  # placeholder class
-                capabilities=caps,
-                metadata={"description": desc, "requires": ["jvm"]},
-                invoke=invoke_fn,
-            )
-        except Exception as e:
-            logger.debug(f"Tweety handler {name} registration deferred: {e}")
-            # Fall back to slot declaration if JVM not available
-            for cap in caps:
-                registry.declare_slot(cap, requires=["jvm"], description=desc)
-
-
-# --- Pre-built workflow definitions ---
-
-
-def build_light_workflow() -> WorkflowDefinition:
-    """Minimal 3-phase analysis workflow with fact extraction."""
-    return (
-        WorkflowBuilder("light_analysis")
-        .add_phase("extract", capability="fact_extraction")
-        .add_phase("quality", capability="argument_quality", depends_on=["extract"])
-        .add_phase(
-            "counter",
-            capability="counter_argument_generation",
-            depends_on=["quality"],
-        )
-        .build()
-    )
-
-
-def build_standard_workflow() -> WorkflowDefinition:
-    """Standard workflow with fact extraction, fallacy detection, and quality-gated counter-arguments.
-
-    Self-hosted LLM (Tier 2) and hierarchical fallacy detection run as optional
-    phases after extraction (#297). Downstream phases (quality, counter,
-    JTMS) read fallacy results via context['phase_hierarchical_fallacy_output'].
-    """
-    return (
-        WorkflowBuilder("standard_analysis")
-        .add_phase("extract", capability="fact_extraction")
-        .add_phase(
-            "neural_detect",
-            capability="neural_fallacy_detection",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .add_phase(
-            "hierarchical_fallacy",
-            capability="hierarchical_fallacy_detection",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .add_phase(
-            "nl_to_logic",
-            capability="nl_to_logic_translation",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .add_phase(
-            "pl",
-            capability="propositional_logic",
-            depends_on=["nl_to_logic"],
-            optional=True,
-        )
-        .add_phase(
-            "fol",
-            capability="fol_reasoning",
-            depends_on=["nl_to_logic"],
-            optional=True,
-        )
-        .add_phase("quality", capability="argument_quality", depends_on=["extract"])
-        .add_phase(
-            "counter",
-            capability="counter_argument_generation",
-            depends_on=["quality"],
-        )
-        .add_phase(
-            "jtms",
-            capability="belief_maintenance",
-            depends_on=["counter"],
-            optional=True,
-        )
-        .add_phase(
-            "governance",
-            capability="governance_simulation",
-            depends_on=["quality"],
-            optional=True,
-        )
-        .add_phase(
-            "debate",
-            capability="adversarial_debate",
-            depends_on=["counter"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-def build_full_workflow() -> WorkflowDefinition:
-    """Full pipeline traversing all 12 capabilities with LLM extraction."""
-    return (
-        WorkflowBuilder("full_analysis")
-        .add_phase(
-            "transcribe",
-            capability="speech_transcription",
-            optional=True,
-        )
-        .add_phase("extract", capability="fact_extraction")
-        .add_phase("quality", capability="argument_quality", depends_on=["extract"])
-        .add_phase(
-            "neural_fallacy",
-            capability="neural_fallacy_detection",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .add_phase(
-            "hierarchical_fallacy",
-            capability="hierarchical_fallacy_detection",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .add_phase(
-            "counter",
-            capability="counter_argument_generation",
-            depends_on=["quality"],
-        )
-        .add_phase(
-            "jtms",
-            capability="belief_maintenance",
-            depends_on=["counter"],
-            optional=True,
-        )
-        .add_phase(
-            "governance",
-            capability="governance_simulation",
-            depends_on=["quality"],
-            optional=True,
-        )
-        .add_phase(
-            "debate",
-            capability="adversarial_debate",
-            depends_on=["counter"],
-            optional=True,
-        )
-        .add_phase(
-            "index",
-            capability="semantic_indexing",
-            depends_on=["counter"],
-            optional=True,
-        )
-        .add_phase(
-            "nl_to_logic",
-            capability="nl_to_logic_translation",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .add_phase(
-            "pl",
-            capability="propositional_logic",
-            depends_on=["nl_to_logic"],
-            optional=True,
-        )
-        .add_phase(
-            "fol",
-            capability="fol_reasoning",
-            depends_on=["nl_to_logic"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-def _should_rerun_fallacy(context: Dict[str, Any]) -> bool:
-    """Check whether the fallacy detection phase should be re-run after JTMS.
-
-    Returns True when JTMS retracted beliefs AND we haven't exceeded the
-    maximum number of fallacy reruns (2). This prevents infinite loops while
-    allowing iterative enrichment when formal analysis invalidates beliefs.
-
-    Args:
-        context: Workflow execution context containing phase outputs.
-
-    Returns:
-        True if a fallacy re-run is warranted.
-    """
-    # Check rerun count limit
-    rerun_count = context.get("_fallacy_rerun_count", 0)
-    if rerun_count >= 2:
-        return False
-
-    # Check if JTMS phase produced retracted beliefs
-    jtms_output = context.get("phase_jtms_output", {})
-    if not isinstance(jtms_output, dict):
-        return False
-
-    # Method 1: explicit undermined_count field
-    undermined = jtms_output.get("undermined_count", 0)
-    if undermined > 0:
-        return True
-
-    # Method 2: scan beliefs dict for invalid entries
-    beliefs = jtms_output.get("beliefs", {})
-    if isinstance(beliefs, dict):
-        retracted = [name for name, bdata in beliefs.items()
-                     if isinstance(bdata, dict) and bdata.get("valid") is False]
-        if retracted:
-            return True
-
-    return False
-
-
-def _increment_fallacy_rerun(context: Dict[str, Any]) -> None:
-    """Increment the fallacy rerun counter in context."""
-    context["_fallacy_rerun_count"] = context.get("_fallacy_rerun_count", 0) + 1
-
-
-def build_iterative_analysis_workflow() -> WorkflowDefinition:
-    """Iterative workflow with JTMS-driven fallacy re-analysis (#305).
-
-    Runs the standard pipeline phases, then conditionally re-runs
-    hierarchical_fallacy after JTMS if beliefs were retracted. This
-    implements the PM re-prompting mechanism for the sequential pipeline:
-    formal analysis findings feed back into informal analysis.
-
-    Flow:
-        extract → hierarchical_fallacy → quality → counter → jtms
-        → [if retractions] fallacy_reanalysis → governance → debate
-    """
-
-    def _fallacy_rerun_condition(ctx: Dict[str, Any]) -> bool:
-        """Condition for the fallacy re-analysis phase."""
-        should_rerun = _should_rerun_fallacy(ctx)
-        if should_rerun:
-            _increment_fallacy_rerun(ctx)
-            logger.info(
-                "Iterative workflow: JTMS retracted beliefs detected, "
-                "re-running fallacy detection (rerun #%d)",
-                ctx.get("_fallacy_rerun_count", 1),
-            )
-        return should_rerun
-
-    return (
-        WorkflowBuilder("iterative_analysis")
-        # Phase 1: extraction
-        .add_phase("extract", capability="fact_extraction")
-        # Phase 2: initial fallacy detection
-        .add_phase(
-            "hierarchical_fallacy",
-            capability="hierarchical_fallacy_detection",
-            depends_on=["extract"],
-            optional=True,
-        )
-        # Phase 3: quality evaluation
-        .add_phase("quality", capability="argument_quality", depends_on=["extract"])
-        # Phase 4: counter-argument generation
-        .add_phase(
-            "counter",
-            capability="counter_argument_generation",
-            depends_on=["quality"],
-        )
-        # Phase 5: JTMS belief maintenance
-        .add_phase(
-            "jtms",
-            capability="belief_maintenance",
-            depends_on=["counter"],
-            optional=True,
-        )
-        # Phase 6: CONDITIONAL fallacy re-analysis after JTMS
-        .add_conditional_phase(
-            "fallacy_reanalysis",
-            capability="hierarchical_fallacy_detection",
-            condition=_fallacy_rerun_condition,
-            depends_on=["jtms"],
-            optional=True,
-        )
-        # Phase 7: governance (depends on quality + optional reanalysis)
-        .add_phase(
-            "governance",
-            capability="governance_simulation",
-            depends_on=["quality"],
-            optional=True,
-        )
-        # Phase 8: debate
-        .add_phase(
-            "debate",
-            capability="adversarial_debate",
-            depends_on=["counter"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-def build_nl_to_logic_workflow() -> WorkflowDefinition:
-    """NL-to-formal-logic pipeline: extract → translate → validate → reason (#173).
-
-    Bridges informal NL analysis and formal reasoning by translating
-    extracted arguments into propositional/FOL formulas with Tweety validation.
-    """
-    return (
-        WorkflowBuilder("nl_to_logic_analysis")
-        .add_phase("extract", capability="fact_extraction")
-        .add_phase(
-            "nl_to_logic",
-            capability="nl_to_logic_translation",
-            depends_on=["extract"],
-        )
-        .add_phase(
-            "pl",
-            capability="propositional_logic",
-            depends_on=["nl_to_logic"],
-            optional=True,
-        )
-        .add_phase(
-            "fol",
-            capability="fol_reasoning",
-            depends_on=["nl_to_logic"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-def build_quality_gated_counter_workflow() -> WorkflowDefinition:
-    """Quality-gated counter-argument with iterative refinement (Loop 3)."""
-
-    def quality_gate(ctx):
-        """Only generate counter-argument if quality score > 3.0."""
-        output = ctx.get("phase_quality_output")
-        if not output or not isinstance(output, dict):
-            return True  # proceed if no quality data
-        return output.get("note_finale", 0) > 3.0
-
-    def counter_quality_convergence(prev, curr):
-        """Converge when counter-argument quality stops improving."""
-        if not isinstance(prev, dict) or not isinstance(curr, dict):
-            return False
-        prev_score = prev.get("note_finale", 0)
-        curr_score = curr.get("note_finale", 0)
-        return curr_score >= prev_score  # stop when no improvement
-
-    return (
-        WorkflowBuilder("quality_gated_counter")
-        .add_phase("extract", capability="fact_extraction")
-        .add_phase("quality", capability="argument_quality", depends_on=["extract"])
-        .add_conditional_phase(
-            "counter",
-            capability="counter_argument_generation",
-            condition=quality_gate,
-            depends_on=["quality"],
-        )
-        .add_loop(
-            "quality_recheck",
-            capability="argument_quality",
-            max_iterations=3,
-            convergence_fn=counter_quality_convergence,
-            depends_on=["counter"],
-            input_transform=lambda inp, ctx: str(
-                ctx.get("phase_counter_output", {})
-                .get("suggested_strategy", {})
-                .get("strategy_name", inp)
-            ),
-        )
-        .build()
-    )
-
-
-def build_debate_governance_loop_workflow() -> WorkflowDefinition:
-    """Debate-Governance vote-contest-debate-revote loop (Loop 1). STUB."""
-    return (
-        WorkflowBuilder("debate_governance_loop")
-        .add_phase(
-            "governance_vote",
-            capability="governance_simulation",
-            optional=True,
-        )
-        .add_phase(
-            "debate",
-            capability="adversarial_debate",
-            depends_on=["governance_vote"],
-            optional=True,
-        )
-        .add_phase(
-            "governance_revote",
-            capability="governance_simulation",
-            depends_on=["debate"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-def build_jtms_dung_loop_workflow() -> WorkflowDefinition:
-    """JTMS-Dung belief retraction/extension recalc loop (Loop 2). STUB."""
-    return (
-        WorkflowBuilder("jtms_dung_loop")
-        .add_phase(
-            "jtms_beliefs",
-            capability="belief_maintenance",
-            optional=True,
-        )
-        .add_phase(
-            "dung_extensions",
-            capability="ranking_semantics",
-            depends_on=["jtms_beliefs"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-def build_neural_symbolic_fallacy_workflow() -> WorkflowDefinition:
-    """Neural-Symbolic-Hierarchical fallacy fusion (Loop 4).
-
-    Three complementary fallacy detection approaches:
-    - Neural: Self-hosted LLM with function calling (#297, replaces CamemBERT)
-    - Hierarchical: Taxonomy-guided iterative deepening (precise, explainable)
-    - Quality baseline: argument quality evaluation for context
-    """
-    return (
-        WorkflowBuilder("neural_symbolic_fallacy")
-        .add_phase(
-            "neural_detect",
-            capability="neural_fallacy_detection",
-            optional=True,
-        )
-        .add_phase(
-            "hierarchical_detect",
-            capability="hierarchical_fallacy_detection",
-            optional=True,
-        )
-        .add_phase("quality_baseline", capability="argument_quality")
-        .build()
-    )
-
-
-def build_hierarchical_fallacy_workflow() -> WorkflowDefinition:
-    """Hierarchical taxonomy-guided fallacy detection workflow (#84).
-
-    Dedicated workflow for the master-slave iterative deepening approach:
-    1. Extract facts/arguments (provides context for fallacy detection)
-    2. Hierarchical fallacy detection via taxonomy navigation
-    3. Quality evaluation for cross-validation
-    """
-    return (
-        WorkflowBuilder("hierarchical_fallacy")
-        .add_phase("extract", capability="fact_extraction")
-        .add_phase(
-            "hierarchical_fallacy",
-            capability="hierarchical_fallacy_detection",
-            depends_on=["extract"],
-        )
-        .add_phase(
-            "quality",
-            capability="argument_quality",
-            depends_on=["extract"],
-            optional=True,
-        )
-        .build()
-    )
-
-
-WORKFLOW_CATALOG: Dict[str, WorkflowDefinition] = {}
-
-
-def get_workflow_catalog() -> Dict[str, WorkflowDefinition]:
-    """Get the catalog of pre-built workflows (lazy-initialized)."""
-    global WORKFLOW_CATALOG
-    if not WORKFLOW_CATALOG:
-        WORKFLOW_CATALOG = {
-            "light": build_light_workflow(),
-            "standard": build_standard_workflow(),
-            "full": build_full_workflow(),
-            "iterative": build_iterative_analysis_workflow(),
-            "quality_gated": build_quality_gated_counter_workflow(),
-            "debate_governance": build_debate_governance_loop_workflow(),
-            "jtms_dung": build_jtms_dung_loop_workflow(),
-            "neural_symbolic": build_neural_symbolic_fallacy_workflow(),
-            "hierarchical_fallacy": build_hierarchical_fallacy_workflow(),
-            "nl_to_logic": build_nl_to_logic_workflow(),
-        }
-        # Collaborative multi-agent debate (#175)
-        try:
-            from argumentation_analysis.orchestration.collaborative_debate import (
-                build_collaborative_analysis_workflow,
-            )
-
-            WORKFLOW_CATALOG["collaborative"] = build_collaborative_analysis_workflow()
-        except Exception as e:
-            logger.warning(f"Collaborative workflow not registered: {e}")
-        # Macro workflows (Track D)
-        try:
-            from argumentation_analysis.workflows.democratech import (
-                build_democratech_workflow,
-            )
-            from argumentation_analysis.workflows.debate_tournament import (
-                build_debate_tournament_workflow,
-            )
-            from argumentation_analysis.workflows.fact_check_pipeline import (
-                build_fact_check_workflow,
-            )
-
-            WORKFLOW_CATALOG["democratech"] = build_democratech_workflow()
-            WORKFLOW_CATALOG["debate_tournament"] = build_debate_tournament_workflow()
-            WORKFLOW_CATALOG["fact_check"] = build_fact_check_workflow()
-        except Exception as e:
-            logger.warning(f"Macro workflows not registered: {e}")
-        # Formal workflows (Track A plugins + orchestration)
-        try:
-            from argumentation_analysis.workflows.formal_debate import (
-                build_formal_debate_workflow,
-            )
-            from argumentation_analysis.workflows.belief_dynamics import (
-                build_belief_dynamics_workflow,
-            )
-            from argumentation_analysis.workflows.argument_strength import (
-                build_argument_strength_workflow,
-            )
-
-            WORKFLOW_CATALOG["formal_debate"] = build_formal_debate_workflow()
-            WORKFLOW_CATALOG["belief_dynamics"] = build_belief_dynamics_workflow()
-            WORKFLOW_CATALOG["argument_strength"] = build_argument_strength_workflow()
-        except Exception as e:
-            logger.warning(f"Formal workflows not registered: {e}")
-        # Formal verification pipeline (#71)
-        try:
-            from argumentation_analysis.workflows.formal_verification import (
-                build_formal_verification_workflow,
-            )
-
-            WORKFLOW_CATALOG[
-                "formal_verification"
-            ] = build_formal_verification_workflow()
-        except Exception as e:
-            logger.warning(f"Formal verification workflow not registered: {e}")
-        # Comprehensive analysis (LLM-only, benchmark-optimized)
-        try:
-            from argumentation_analysis.workflows.comprehensive_analysis import (
-                build_comprehensive_analysis_workflow,
-            )
-
-            WORKFLOW_CATALOG["comprehensive"] = build_comprehensive_analysis_workflow()
-        except Exception as e:
-            logger.warning(f"Comprehensive workflow not registered: {e}")
-    return WORKFLOW_CATALOG
-
-
-def reset_workflow_catalog() -> None:
-    """
-    Reset the global workflow catalog to an empty state.
-
-    This is primarily useful for testing to ensure that tests don't pollute
-    each other's state through the global WORKFLOW_CATALOG variable.
-
-    After calling this function, the next call to get_workflow_catalog() will
-    re-initialize the catalog with all available workflows.
-    """
-    global WORKFLOW_CATALOG
-    WORKFLOW_CATALOG = {}
-    logger.debug("Workflow catalog reset to empty state")
-
-
-async def run_unified_analysis(
-    text: str,
-    workflow_name: str = "standard",
-    registry: Optional[CapabilityRegistry] = None,
-    custom_workflow: Optional[WorkflowDefinition] = None,
-    context: Optional[Dict[str, Any]] = None,
-    state: Optional[Any] = None,
-    create_state: bool = True,
-) -> Dict[str, Any]:
-    """
-    Run a unified analysis pipeline on input text.
-
-    Args:
-        text: The text to analyze.
-        workflow_name: Name of pre-built workflow ("light", "standard", "full").
-        registry: CapabilityRegistry to use. If None, creates one via setup_registry().
-        custom_workflow: If provided, use this workflow instead of a named one.
-        context: Additional context passed to each phase (e.g. kernel, config).
-        state: Optional UnifiedAnalysisState instance. If provided, phase results
-               are written to it via state writers.
-        create_state: If True and no state provided, automatically create an
-                      UnifiedAnalysisState. Set to False to disable state tracking.
-
-    Returns:
-        Dict with keys:
-            - workflow_name: Name of the executed workflow
-            - phases: Dict[str, PhaseResult] — per-phase results
-            - summary: Dict with completed/failed/skipped counts
-            - capabilities_used: List of capabilities that were successfully resolved
-            - capabilities_missing: List of capabilities with no provider
-            - unified_state: UnifiedAnalysisState (if state tracking enabled)
-            - state_snapshot: Dict snapshot of the state (if state tracking enabled)
-    """
-    if registry is None:
-        registry = setup_registry()
-
-    # Create or use provided state
-    if state is None and create_state:
-        try:
-            from argumentation_analysis.core.shared_state import UnifiedAnalysisState
-
-            state = UnifiedAnalysisState(text)
-        except ImportError:
-            logger.warning(
-                "Could not import UnifiedAnalysisState; state tracking disabled"
-            )
-            state = None
-
-    if custom_workflow is not None:
-        workflow = custom_workflow
-    elif workflow_name == "auto":
-        try:
-            from argumentation_analysis.orchestration.router import TextAnalysisRouter
-
-            router = TextAnalysisRouter(registry=registry)
-            workflow = await router.analyze_and_route_async(text, registry=registry)
-        except Exception as e:
-            logger.warning("Auto-routing failed, falling back to standard: %s", e)
-            catalog = get_workflow_catalog()
-            workflow = catalog["standard"]
-    elif workflow_name == "conversational":
-        # Special mode: use ConversationalOrchestrator (#208-L)
-        try:
-            from argumentation_analysis.orchestration.conversational_orchestrator import (
-                run_conversational_analysis,
-            )
-
-            conv_result = await run_conversational_analysis(text)
-
-            # Normalize result format for benchmark compatibility (#208-L)
-            # Conversational returns {phases: [names], conversation_log, total_messages}
-            # Benchmark expects {phases: {dict}, summary: {completed, total, ...}}
-            phase_names = conv_result.get("phases", [])
-            total_msgs = conv_result.get("total_messages", 0)
-            conv_result["summary"] = {
-                "completed": len(phase_names),
-                "failed": 0,
-                "skipped": 0,
-                "total": len(phase_names),
-                "total_messages": total_msgs,
-            }
-            conv_result["workflow_name"] = "conversational"
-            return conv_result
-        except Exception as e:
-            logger.warning(
-                f"Conversational mode failed ({e}), falling back to standard"
-            )
-            catalog = get_workflow_catalog()
-            workflow = catalog["standard"]
-    else:
-        catalog = get_workflow_catalog()
-        if workflow_name not in catalog:
-            raise ValueError(
-                f"Unknown workflow '{workflow_name}'. "
-                f"Available: {list(catalog.keys()) + ['conversational']}"
-            )
-        workflow = catalog[workflow_name]
-
-    executor = WorkflowExecutor(registry)
-    phase_results = await executor.execute(
-        workflow,
-        input_data=text,
-        context=context,
-        state=state,
-        state_writers=CAPABILITY_STATE_WRITERS if state is not None else None,
-    )
-
-    # Build summary
-    completed = [
-        name for name, r in phase_results.items() if r.status == PhaseStatus.COMPLETED
-    ]
-    failed = [
-        name for name, r in phase_results.items() if r.status == PhaseStatus.FAILED
-    ]
-    skipped_phases = [
-        name for name, r in phase_results.items() if r.status == PhaseStatus.SKIPPED
-    ]
-
-    capabilities_used = [
-        r.capability
-        for r in phase_results.values()
-        if r.status == PhaseStatus.COMPLETED
-    ]
-    capabilities_missing = [
-        r.capability
-        for r in phase_results.values()
-        if r.status in (PhaseStatus.FAILED, PhaseStatus.SKIPPED)
-    ]
-
-    result = {
-        "workflow_name": workflow.name,
-        "phases": phase_results,
-        "summary": {
-            "completed": len(completed),
-            "failed": len(failed),
-            "skipped": len(skipped_phases),
-            "total": len(phase_results),
-            "completed_phases": completed,
-            "failed_phases": failed,
-            "skipped_phases": skipped_phases,
-        },
-        "capabilities_used": capabilities_used,
-        "capabilities_missing": capabilities_missing,
-    }
-
-    # Include state in results if available
-    if state is not None:
-        result["unified_state"] = state
-        try:
-            result["state_snapshot"] = state.get_state_snapshot(summarize=True)
-        except Exception:
-            result["state_snapshot"] = None
-
-    return result
